AdFisher/requirements.txt
# These packages can be installed by running:
#
#   pip install -r requirements.txt
#
selenium
xvfbwrapper
numpy
scipy
matplotlib
scikit-learn
stemming
pyyaml
nltk
psutil
adblockparser
re2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

